Feat: 커뮤니티 페이지 empty case 추가
