+++ src/app/community/_components/PostCardList.module.scss
@@ -1,9 +1,3 @@
-.filter-write-button-wrapper {
-  display: flex;
-  justify-content: space-between;
-  width: 100%;
-}
-
 .post-wrapper {
   display: grid;
   grid-template-columns: repeat(4, 1fr);
