+++ src/app/community/page.tsx
@@ -1,13 +1,8 @@
-import classNames from 'classnames/bind';
-
 import { getAllCommunityPost } from '@/api/communityAPI';
 import Pagination from '@/components/Pagination/Pagination';
 import type { CommunityParamsType } from '@/types/CommunityTypes';
+import CommunityEmptyCase from './_components/CommunityEmptyCase';
 import PostCardList from './_components/PostCardList';
-
-import styles from './page.module.scss';
-
-const cn = classNames.bind(styles);
 
 interface CommunityPageProps {
   searchParams: { [key: string]: string | undefined };
@@ -22,11 +17,14 @@ export default async function CommunityPage({ searchParams }: CommunityPageProps
 
   const data = await getAllCommunityPost(initialParams);
 
+  if (!data) {
+    return <CommunityEmptyCase />;
+  }
+
   const { content, ...rest } = data;
 
   return (
-    <div className={cn('container')}>
-      <p className={cn('page-name')}>커뮤니티</p>
+    <div>
       <PostCardList searchParams={searchParams} initialData={content} />
       <Pagination {...rest} searchParams={searchParams} />
     </div>
